separated some logics from Project class to new classes GraphQLUtils and GraphQLDocumentCreator
